Bring `hydrogen` back to ns naming pattern
This library is a `hydrogen` module - while moving `hydrogen`-related
libraries to HOP organization, we do not want to get rid of `hydrogen`
name or change it to `dev.gethop` - this is still `hydrogen`.
What we need is to move the repository to `gethop-dev` Github
organization and publish releases under `dev.gethop` group. So
`hydrogen` brand name should remains as it was.